# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gmail Client API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/status": {
      "get": {
        "summary": "Auth Status",
        "description": "Check if user is authenticated (survives browser refresh + app restart).",
        "operationId": "auth_status_api_auth_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "description": "Login with email + app password. Session is persisted to disk.",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Logout",
        "description": "Disconnect, cancel all tasks, clear saved session.",
        "operationId": "logout_api_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/auth/verify-lock": {
      "post": {
        "summary": "Verify Lock",
        "operationId": "verify_lock_api_auth_verify_lock_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnlockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/emails/fetch": {
      "post": {
        "summary": "Start Fetch",
        "description": "Start a full email fetch in the background.",
        "operationId": "start_fetch_api_emails_fetch_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/emails/refresh": {
      "post": {
        "summary": "Start Refresh",
        "description": "Start an incremental fetch (new emails only) in the background.",
        "operationId": "start_refresh_api_emails_refresh_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/emails/delete": {
      "post": {
        "summary": "Start Delete",
        "description": "Delete emails by UIDs in the background.",
        "operationId": "start_delete_api_emails_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/emails/delete-by-sender": {
      "post": {
        "summary": "Start Delete By Sender",
        "description": "Delete all emails from a sender in the background.",
        "operationId": "start_delete_by_sender_api_emails_delete_by_sender_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteBySenderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/emails/delete-bulk-senders": {
      "post": {
        "summary": "Start Delete Bulk Senders",
        "description": "Delete all emails from multiple senders in the background.",
        "operationId": "start_delete_bulk_senders_api_emails_delete_bulk_senders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteSendersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/emails/progress": {
      "get": {
        "summary": "Get Progress",
        "description": "Get current operation progress. Browser polls this.",
        "operationId": "get_progress_api_emails_progress_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/emails": {
      "get": {
        "summary": "Get Emails",
        "description": "Get fetched emails with search, sort, pagination.",
        "operationId": "get_emails_api_emails_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "date",
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "desc",
              "title": "Sort Order"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/emails/senders": {
      "get": {
        "summary": "Get Senders",
        "description": "Get email stats grouped by sender.",
        "operationId": "get_senders_api_emails_senders_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "total",
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "desc",
              "title": "Sort Order"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 30,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/emails/summary": {
      "get": {
        "summary": "Get Summary",
        "operationId": "get_summary_api_emails_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/auto-pull": {
      "get": {
        "summary": "Get Auto Pull",
        "operationId": "get_auto_pull_api_auto_pull_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Set Auto Pull",
        "operationId": "set_auto_pull_api_auto_pull_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutoPullRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AutoPullRequest": {
        "properties": {
          "interval": {
            "type": "integer",
            "title": "Interval"
          }
        },
        "type": "object",
        "required": [
          "interval"
        ],
        "title": "AutoPullRequest"
      },
      "BulkDeleteSendersRequest": {
        "properties": {
          "sender_emails": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sender Emails"
          }
        },
        "type": "object",
        "required": [
          "sender_emails"
        ],
        "title": "BulkDeleteSendersRequest"
      },
      "DeleteBySenderRequest": {
        "properties": {
          "sender_email": {
            "type": "string",
            "title": "Sender Email"
          }
        },
        "type": "object",
        "required": [
          "sender_email"
        ],
        "title": "DeleteBySenderRequest"
      },
      "DeleteRequest": {
        "properties": {
          "uids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Uids"
          }
        },
        "type": "object",
        "required": [
          "uids"
        ],
        "title": "DeleteRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "app_password": {
            "type": "string",
            "title": "App Password"
          },
          "lock_code": {
            "type": "string",
            "title": "Lock Code"
          }
        },
        "type": "object",
        "required": [
          "email",
          "app_password",
          "lock_code"
        ],
        "title": "LoginRequest"
      },
      "UnlockRequest": {
        "properties": {
          "lock_code": {
            "type": "string",
            "title": "Lock Code"
          }
        },
        "type": "object",
        "required": [
          "lock_code"
        ],
        "title": "UnlockRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}